refactor: add option to keep file names

+++ src/util/cmds/lock.cmd.zig
@@ -4,9 +4,10 @@ const Command = @import("../cmds.zig");
 const write = @import("../color.zig").write;
 const Identity = @import("../../crypto/identity.zig");
 const Box = @import("../sync.zig").Box;
+const Naming = @import("root").Naming;
 
 /// Attempt to encrypt a folder or file at a specified path
-pub fn call(c: *Command, path: []const u8) Command.Error!u8 {
+pub fn call(c: *Command, n: Naming, path: []const u8) Command.Error!u8 {
     std.fs.cwd().access(path, .{}) catch {
         try write(c.stderr.writer(), .Red, "error:");
         try write(c.stderr.writer(), .Default, " ");
@@ -17,7 +18,7 @@ pub fn call(c: *Command, path: []const u8) Command.Error!u8 {
     const info = try std.fs.cwd().statFile(path);
 
     switch (info.kind) {
-        .file => if (try encryptFile(c, path)) |rename| switch (rename) {
+        .file => if (try encryptFile(c, n, path)) |rename| switch (rename) {
             .changed => |meta| {
                 try write(c.stdout.writer(), .Green, "successfully ");
                 try write(c.stdout.writer(), .Default, "encrypted ");
@@ -40,7 +41,7 @@ pub fn call(c: *Command, path: []const u8) Command.Error!u8 {
             try c.stderr.writeAll("the file is already encrypted\n");
             return 1;
         },
-        .directory => switch (try encryptDir(c, path)) {
+        .directory => switch (try encryptDir(c, n, path)) {
             .changed => |meta| {
                 try write(c.stdout.writer(), .Green, "successfully ");
                 try write(c.stdout.writer(), .Default, "encrypted ");
@@ -68,20 +69,17 @@ pub fn call(c: *Command, path: []const u8) Command.Error!u8 {
 }
 
 // Attempt to encrypt a file at the given path
-fn encryptFile(c: *Command, path: []const u8) Command.Error!?sys.Rename {
+fn encryptFile(c: *Command, n: Naming, path: []const u8) Command.Error!?sys.Rename {
     var id = try Identity.load(try c.keys.read(.secret));
     const file = try sys.File.load(c.allocator, path);
     defer file.unload(c.allocator);
 
     if (std.mem.eql(u8, file.data[0..5], &sys.magic)) return null;
 
-    const name = try c.allocator.dupe(u8, file.meta.name);
-    errdefer c.allocator.free(name);
+    const old_name = try c.allocator.dupe(u8, file.meta.name);
+    errdefer c.allocator.free(old_name);
 
-    const hash = try c.crypt.hash(&id, file.data);
-    errdefer c.allocator.free(hash);
-
-    const cdata = try std.mem.concat(c.allocator, u8, &.{ name, &.{1}, file.data });
+    const cdata = try std.mem.concat(c.allocator, u8, &.{ old_name, &.{1}, file.data });
     defer c.allocator.free(cdata);
 
     const data = try c.crypt.encrypt(&id, cdata, .b64);
@@ -96,14 +94,25 @@ fn encryptFile(c: *Command, path: []const u8) Command.Error!?sys.Rename {
     var dir = try std.fs.openDirAbsolute(file.meta.dir, .{});
     defer dir.close();
 
-    try dir.writeFile(.{ .sub_path = hash, .data = sdata });
-    try dir.deleteFile(file.meta.path);
+    if (n == .change) {
+        const new_name = try c.crypt.hash(&id, file.data);
+        errdefer c.allocator.free(new_name);
 
-    return sys.Rename{ .changed = .{ .old = name, .new = hash } };
+        try dir.writeFile(.{ .sub_path = new_name, .data = sdata });
+        try dir.deleteFile(file.meta.path);
+
+        return .{ .changed = .{
+            .old = old_name,
+            .new = new_name,
+        } };
+    }
+
+    try dir.writeFile(.{ .sub_path = old_name, .data = sdata });
+    return .{ .kept = old_name };
 }
 
 // Attempt to encrypt a directory at the given path
-fn encryptDir(c: *Command, path: []const u8) Command.Error!sys.Rename {
+fn encryptDir(c: *Command, n: Naming, path: []const u8) Command.Error!sys.Rename {
     var id = try Identity.load(try c.keys.read(.secret));
     var dir = try std.fs.cwd().openDir(path, .{ .iterate = true });
     defer dir.close();
@@ -112,11 +121,11 @@ fn encryptDir(c: *Command, path: []const u8) Command.Error!sys.Rename {
     var iterator = Box(std.fs.Dir.Iterator).init(dir.iterate());
 
     if (c.threads.allowed == 1) {
-        try worker(c, &dir, &group, &iterator);
+        try worker(c, &dir, &group, &iterator, n);
     } else {
         c.mutex.lock();
         while (c.threads.used < c.threads.allowed) : (c.threads.used += 1) {
-            _ = try std.Thread.spawn(.{}, worker, .{ c, &dir, &group, &iterator });
+            _ = try std.Thread.spawn(.{}, worker, .{ c, &dir, &group, &iterator, n });
         }
         c.mutex.unlock();
     }
@@ -129,16 +138,23 @@ fn encryptDir(c: *Command, path: []const u8) Command.Error!sys.Rename {
     const old_name = try c.allocator.dupe(u8, name_buf);
     errdefer c.allocator.free(old_name);
 
-    const parent = std.fs.path.dirname(rpath);
-    const new_name = try c.crypt.encrypt(&id, old_name, .hex);
-    errdefer c.allocator.free(new_name);
+    if (n == .change) {
+        const parent = std.fs.path.dirname(rpath);
+        const new_name = try c.crypt.encrypt(&id, old_name, .hex);
+        errdefer c.allocator.free(new_name);
 
-    const npath = try std.fs.path.join(c.allocator, &.{ parent.?, new_name });
-    defer c.allocator.free(npath);
+        const npath = try std.fs.path.join(c.allocator, &.{ parent.?, new_name });
+        defer c.allocator.free(npath);
 
-    try std.fs.renameAbsolute(rpath, npath);
+        try std.fs.renameAbsolute(rpath, npath);
 
-    return sys.Rename{ .changed = .{ .old = old_name, .new = new_name } };
+        return .{ .changed = .{
+            .old = old_name,
+            .new = new_name,
+        } };
+    }
+
+    return .{ .kept = old_name };
 }
 
 fn worker(
@@ -146,6 +162,7 @@ fn worker(
     d: *std.fs.Dir,
     wg: *std.Thread.WaitGroup,
     it: *Box(std.fs.Dir.Iterator),
+    n: Naming,
 ) Command.Error!void {
     wg.start();
 
@@ -162,7 +179,7 @@ fn worker(
         .directory => {
             const sub_path = try d.realpathAlloc(c.allocator, entry.name);
             defer c.allocator.free(sub_path);
-            switch (try encryptDir(c, sub_path)) {
+            switch (try encryptDir(c, n, sub_path)) {
                 .kept => |name| c.allocator.free(name),
                 .changed => |meta| {
                     c.allocator.free(meta.old);
@@ -173,8 +190,7 @@ fn worker(
         .file => {
             const sub_path = try d.realpathAlloc(c.allocator, entry.name);
             defer c.allocator.free(sub_path);
-            const rename = try encryptFile(c, sub_path);
-            if (rename) |r| switch (r) {
+            if (try encryptFile(c, n, sub_path)) |r| switch (r) {
                 .kept => |name| c.allocator.free(name),
                 .changed => |meta| {
                     c.allocator.free(meta.old);
